requirements.txt
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1
rich==13.7.0
alive-progress==3.1.5
art==6.1
